## [3.0.0] - 2020-05-13 - installing packages via tag with opts based on tag

+++ setup.ps1
@@ -1,20 +1,24 @@
 param(
     [parameter(Mandatory = $true)]
-    [String]$tag
+    [String]$tag,
+    [parameter(Mandatory = $false)]
+    [String]$dry
 )
 
 $userHome = Resolve-Path "~"
 
 $paths = @(
-    "$userHome/Google Drive/choco/packages.json",
-    "$userHome/Dropbox/choco/packages.json",
-    "$userHome/choco/packages.json"
+"$userHome/Google Drive/choco/packages.json",
+"$userHome/Dropbox/choco/packages.json",
+"$userHome/choco/packages.json"
 )
 
-foreach ($packageJson in $paths) {
+foreach ($packageJson in $paths)
+{
     Write-Host $packageJson
 
-    if([System.IO.File]::Exists($packageJson)) {
+    if ( [System.IO.File]::Exists($packageJson))
+    {
         break
     }
 }
@@ -27,21 +31,21 @@ $packagesToInstallWithOpts = @()
 
 foreach ($package in $packages)
 {
-    if ($package.tags.count -eq 0 -or $package.tags -notcontains $tag)
+    if ($package.tags -notmatch $tag)
     {
         $packagesToDelete += $package.name
     }
 
-    if ($package.tags.count -gt 0 -and $package.tags -contains $tag)
+    if ($package.tags -match $tag)
     {
-        if (-not([string]::IsNullOrEmpty($package.opts)))
+        if (-not([string]::IsNullOrEmpty($package.tags.$tag)))
         {
             $packageName = $package.name
-            $packageOptions = $package.opts
+            $packageOptions = $package.tags.$tag
             $packagesToInstallWithOpts += "$packageName $packageOptions"
         }
 
-        if (([string]::IsNullOrEmpty($package.opts)))
+        if (([string]::IsNullOrEmpty($package.tags.$tag)))
         {
             $packagesToInstall += $package.name
         }
@@ -52,14 +56,38 @@ $packagesToDelete = $packagesToDelete | Sort-Object
 $packagesToInstall = $packagesToInstall | Sort-Object
 $packagesToInstallWithOpts = $packagesToInstallWithOpts | Sort-Object
 
-$command = "choco uninstall --yes $packagesToDelete"
-Invoke-Expression $command
-
-$command = "choco upgrade --yes $packagesToInstall"
-Invoke-Expression $command
-
-foreach ($package in $packagesToInstallWithOpts)
+if (-not([string]::IsNullOrEmpty($packagesToDelete)))
 {
-    $command = "choco install --yes $package"
-    Invoke-Expression $command
+    $command = "choco uninstall --yes $packagesToDelete"
+    Write-Host $command
+
+    if (([string]::IsNullOrEmpty($dry)))
+    {
+        Invoke-Expression $command
+    }
+}
+
+if (-not([string]::IsNullOrEmpty($packagesToInstall)))
+{
+    $command = "choco upgrade --yes $packagesToInstall"
+    Write-Host $command
+
+    if (([string]::IsNullOrEmpty($dry)))
+    {
+        Invoke-Expression $command
+    }
+}
+
+if (-not([string]::IsNullOrEmpty($packagesToInstallWithOpts)))
+{
+    foreach ($package in $packagesToInstallWithOpts)
+    {
+        $command = "choco install --yes $package"
+        Write-Host $command
+
+        if (([string]::IsNullOrEmpty($dry)))
+        {
+            Invoke-Expression $command
+        }
+    }
 }
